+++ components/shared/sectionHeading.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 
-const SectionHeading = ({heading}:string | any) => {
+const SectionHeading = ({ heading }: string | any) => {
   return (
-      <h1 className='uppercase text-2xl font-extrabold mb-10 xxs:mb-2 xs:mb-5 text-center'>{heading}</h1>
+    <h1 className='uppercase text-2xl font-extrabold mb-10 xxs:mb-2 xs:mb-5 text-center first-letter:text-4xl first-letter:text-[#00d4ff] leading-6 tracking-wider'>{heading}</h1>
   )
 }